SRC: Download in a expansion folder and check before download

+++ src/scrapeImages.py
@@ -4,8 +4,9 @@
 from bs4 import BeautifulSoup
 import urllib
 import requests
+import os
 
-#This fetches the first card from the page
+# This fetches the first card from the page
 def getFirstPic(url):
     result = requests.get(url)
     html = result.content
@@ -15,7 +16,8 @@ def getFirstPic(url):
         return hold.a.get_text()
     else:
         return ""
-#This generates a list of xml samples of card titles
+
+# This generates a list of xml samples of card titles
 def getPics(url):
     result = requests.get(url)
     html = result.content
@@ -23,29 +25,43 @@ def getPics(url):
     samples = soup.find_all("span", "cardTitle")
     return samples
 
-#Gets the card name and multiverseID's from a sample and then saves them
+# Gets the card name and multiverseID's from a sample and then saves them
 def downloadPic(i, samples, set):
-    multiverseID= samples[i].a.attrs['href']
-    sampleName=samples[i].a.get_text()
-    print(sampleName)
-    sampleName=sampleName.replace('/','||')
-    cardName = "data/images/"+sampleName+"   "+set[:-1].replace('%20',' ')+".jpg"
-    URL = "http://gatherer.wizards.com/Handlers/Image.ashx?multiverseid=%s&type=card" %multiverseID[34:]
-    urllib.urlretrieve(URL, cardName)
 
-#This loops through each set in cardSets.txt
+    expansion = 'data/images/'+set[:-1].replace('%20', '_')
+
+    if not os.path.exists(expansion):
+        os.makedirs(expansion)
+
+    multiverseID = samples[i].a.attrs['href']
+    multiverseID = multiverseID[34:]
+
+    sampleName = samples[i].a.get_text()
+    sampleName = sampleName.replace('/', '||')
+
+    print("- "+sampleName)
+
+    cardName = expansion+"/"+multiverseID+".jpg"
+    URL = "http://gatherer.wizards.com/Handlers/Image.ashx?multiverseid=%s&type=card" %multiverseID
+
+    if not os.path.exists(cardName):
+        urllib.urlretrieve(URL, cardName)
+
+# This loops through each set in cardSets.txt
 with open('config/cardSets.txt') as sets:
     samples = []
     lastFirstPic = ""
     for line in sets:
         n=0
         while(n<5):
-            #This loops through the different pages page=0, page=1 etc.
+
+            # This loops through the different pages page=0, page=1 etc.
             URL = "http://gatherer.wizards.com/Pages/Search/Default.aspx?page="+str(n)+"&set=%5B\""+line[:-1]+"\"%5D"
-            #Check if the first pic is the same as the last first pic, if it is not download the images, or else exit the loop
+
+            # Check if the first pic is the same as the last first pic, if it is not download the images, or else exit the loop
             firstPic=getFirstPic(URL)
             if(firstPic!=lastFirstPic):
-                print("-----------------------------------------------")
+                print("\n-----------------------------------------------")
                 print("Downloading images from "+line[:-1].replace('%20',' '))
                 print("-----------------------------------------------")
                 samples = getPics(URL)
